Refactor to group data type definitions in 1 globally accessed file
